refactor(utxo): unify Utxo struct + widen amount to u128 (#2287)
Step 1 of 0.A Codebase hygiene: collapse the duplicate Utxo / OutPoint /
UtxoMerkleProof definitions into a single canonical home and widen the
value field to match 18-decimal SOV.

Type unification:
- Move Utxo, OutPoint, UtxoMerkleProof to lib-types::primitives — single
  source of truth, single serialization, no drift.
- lib-blockchain::storage now re-exports from lib-types.
- lib-utxo collapsed to a thin facade that re-exports the canonical
  types; deleted the duplicate types.rs / apply.rs / errors.rs (the
  apply-layer hardening will be ported in a follow-up — see #2287).

Value-domain widening (Utxo.amount: u64 → u128):
- u64 caps a single UTXO at ~18.4 whole SOV with 18 decimals; widen to
  Amount (u128) to match the rest of the conservation chain.
- Propagate u128 through apply_native_transfer + apply_coinbase, the
  TransferOutcome / CoinbaseOutcome carriers, and the ValueMismatch /
  InsufficientInputs error variants.
- ApplyOutcome.fees_collected widened to u128 (executor) and call sites
  drop the now-redundant `as u64` downcast on FeeModelV2.block_reward.
- Test assertions in phase3c_fee_distribution_tests cast u64 fee
  literals to u128 at the comparison boundary.

Transaction.fee remains u64 — a temporary `as u128` cast bridges the
conservation math at the validation/apply boundary. Follow-up tracked
in #2290.

+++ lib-blockchain/src/execution/errors.rs
@@ -98,10 +98,14 @@ pub enum TxApplyError {
     UtxoAlreadySpent(OutPoint),
 
     #[error("Input/output value mismatch: inputs={inputs}, outputs={outputs}, fee={fee}")]
-    ValueMismatch { inputs: u64, outputs: u64, fee: u64 },
+    ValueMismatch {
+        inputs: u128,
+        outputs: u128,
+        fee: u128,
+    },
 
     #[error("Insufficient input value: have {have}, need {need}")]
-    InsufficientInputs { have: u64, need: u64 },
+    InsufficientInputs { have: u128, need: u128 },
 
     // =========================================================================
     // Token Errors (stateful)
