+++ lib-blockchain/src/validation/errors.rs
@@ -112,7 +112,7 @@ pub enum TxValidateError {
     UtxoNotFound(OutPoint),
 
     #[error("Insufficient inputs: have {have}, need {need}")]
-    InsufficientInputs { have: u64, need: u64 },
+    InsufficientInputs { have: u128, need: u128 },
 
     // =========================================================================
     // Fee Errors
